Various improvements (#94)
* Fix cartpole example. Update to Pinocchio 2.4.3 for manyinux wheel.
* Add time to the observation dict. Fix _update_observation. Varous minor fixes.
* Enable to automatically detect the most appropriate contact frame when computing the freeflyer.
* Add more dynamics helpers and fix existing ones.

+++ examples/cartpole_py/test_simulation.py
@@ -47,7 +47,7 @@ robot.set_options(robot_options)
 engine_py.set_engine_options(engine_options)
 engine_py.set_controller_options(ctrl_options)
 
-engine_py.seed(0)
+engine_py.seed(np.uint32(0))
 engine_py.reset(np.zeros(robot.nx))
 
 refresh_time_prev = time.time()
